+++ src/solace_ai_connector/common/messaging/solace_messaging.py
@@ -250,13 +250,14 @@ class SolaceMessaging(Messaging):
         broker_message = self.persistent_receivers[0].receive_message(timeout_ms)
         if broker_message is None:
             return None
-        
+
         # Convert Solace message to dictionary format
         return {
-            'payload': broker_message.get_payload_as_bytes(),
-            'topic': broker_message.get_destination_name(),
-            'user_properties': broker_message.get_properties(),
-            '_original_message': broker_message  # Keep original message for acknowledgement
+            "payload": broker_message.get_payload_as_string()
+            or broker_message.get_payload_as_bytes(),
+            "topic": broker_message.get_destination_name(),
+            "user_properties": broker_message.get_properties(),
+            "_original_message": broker_message,  # Keep original message for acknowledgement
         }
 
     def subscribe(
@@ -266,7 +267,7 @@ class SolaceMessaging(Messaging):
         persistent_receiver.add_subscription(sub)
 
     def ack_message(self, broker_message):
-        if '_original_message' in broker_message:
-            self.persistent_receiver.ack(broker_message['_original_message'])
+        if "_original_message" in broker_message:
+            self.persistent_receiver.ack(broker_message["_original_message"])
         else:
             log.warning("Cannot acknowledge message: original Solace message not found")
