+++ src/solace_ai_connector/components/inputs_outputs/broker_input.py
@@ -110,16 +110,18 @@ class BrokerInput(BrokerBase):
     def get_next_message(self, timeout_ms=None):
         if timeout_ms is None:
             timeout_ms = DEFAULT_TIMEOUT_MS
-        broker_message = self.messaging_service.receive_message(timeout_ms, self.broker_properties["queue_name"])
+        broker_message = self.messaging_service.receive_message(
+            timeout_ms, self.broker_properties["queue_name"]
+        )
         if not broker_message:
             return None
         self.current_broker_message = broker_message
 
-        payload = broker_message.get('payload')
+        payload = broker_message.get("payload")
         payload = self.decode_payload(payload)
 
-        topic = broker_message.get('topic')
-        user_properties = broker_message.get('user_properties', {})
+        topic = broker_message.get("topic")
+        user_properties = broker_message.get("user_properties", {})
         log.debug(
             "Received message from broker: topic=%s, user_properties=%s, payload length=%d",
             topic,
